refactor: Http11Processor의 Session 관리 로직 SessionSupport 클래스로 분리
